Fix logfile creation.
If etmtk_log.txt does not already exist, etm would try to create it, but would
fail with NameErrors on fp and dfile_encoding.  Since encoding doesn't matter
when we're just creating an empty file, replace it with a short call to open.
Also note in the README that this file is created as part of initialization.